+++ src/NET/Demos/Demo.CustomFolderBrowserDialog/CustomFrameworkDialogFactory.cs
@@ -1,0 +1,13 @@
+﻿using MvvmDialogs.FrameworkDialogs;
+using MvvmDialogs.FrameworkDialogs.FolderBrowser;
+
+namespace Demo.CustomFolderBrowserDialog
+{
+    public class CustomFrameworkDialogFactory : DefaultFrameworkDialogFactory
+    {
+        public override IFrameworkDialog CreateFolderBrowserDialog(FolderBrowserDialogSettings settings)
+        {
+            return new CustomFolderBrowserDialog(settings);
+        }
+    }
+}