+++ src/NET/Demos/Demo.CustomFolderBrowserDialog/MainWindow.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace Demo.CustomFolderBrowserDialog
+{
+    public partial class MainWindow
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+        }
+    }
+}