+++ src/NET/Demos/Demo.ModalDialogTest/ScreenObjects/AddTextScreen.cs
@@ -23,7 +23,7 @@ namespace Demo.ModalDialogTest.ScreenObjects
 
         public virtual string Text
         {
-            set { text.Text = value; }
+            set => text.Text = value;
         }
 
         public virtual void ClickOK() => okButton.Click();