+++ src/NET/Demos/Demo.NonModalCustomDialog/CurrentTimeCustomDialog.cs
@@ -15,20 +15,20 @@ namespace Demo.NonModalCustomDialog
 
         object IWindow.DataContext
         {
-            get { return dialog.DataContext; }
-            set { dialog.DataContext = value; }
+            get => dialog.DataContext;
+            set => dialog.DataContext = value;
         }
 
         bool? IWindow.DialogResult
         {
-            get { return dialog.DialogResult; }
-            set { dialog.DialogResult = value; }
+            get => dialog.DialogResult;
+            set => dialog.DialogResult = value;
         }
 
         ContentControl IWindow.Owner
         {
-            get { return dialog.Owner; }
-            set { dialog.Owner = (Window)value; }
+            get => dialog.Owner;
+            set => dialog.Owner = (Window)value;
         }
 
         bool? IWindow.ShowDialog()