+++ src/NET/Demos/Demo.SaveFileDialog/MainWindowViewModel.cs
@@ -23,7 +23,7 @@ namespace Demo.SaveFileDialog
 
         public string Path
         {
-            get { return path; }
+            get => path;
             private set { Set(() => Path, ref path, value); }
         }