+++ src/Universal/Demo/Demo.PickSingleFolder/MainPageViewModel.cs
@@ -25,7 +25,7 @@ namespace Demo.PickSingleFolder
 
         public string Path
         {
-            get { return path; }
+            get => path;
             private set { Set(() => Path, ref path, value); }
         }